requirements.txt
Django==1.6
South==1.0
amqp==1.4.5
anyjson==0.3.3
argparse==1.2.1
billiard==3.3.0.18
celery==3.1.13
django-celery==3.1.10
djangorestframework==2.3.14
kombu==3.0.21
psycopg2==2.5.3
pytz==2014.4
redis==2.10.1
requests==2.3.0
wsgiref==0.1.2
